chore(trie): reorganise files by technical domains
