+++ packages/trie/src/db/index.ts
@@ -1,0 +1,2 @@
+export * from './checkpoint'
+export * from './db'
