+++ packages/trie/src/index.ts
@@ -1,5 +1,5 @@
-export { CheckpointTrie } from './checkpointTrie'
 export * from './db'
-export { SecureTrie } from './secure'
-export { Trie as BaseTrie } from './baseTrie'
-export { WalkController } from './util/walkController'
+export * from './proof'
+export * from './trie'
+export * from './types'
+export * from './util'
