+++ packages/trie/src/proof/index.ts
@@ -1,0 +1,1 @@
+export * from './range'
