+++ packages/trie/src/trie/node.ts
@@ -1,8 +1,8 @@
 import { keccak256 } from 'ethereum-cryptography/keccak'
 import { arrToBufArr, bufArrToArr } from '@ethereumjs/util'
 import RLP from 'rlp'
-import { bufferToNibbles, nibblesToBuffer } from './util/nibbles'
-import { isTerminator, addHexPrefix, removeHexPrefix } from './util/hex'
+import { bufferToNibbles, nibblesToBuffer } from '../util/nibbles'
+import { isTerminator, addHexPrefix, removeHexPrefix } from '../util/hex'
 
 export type TrieNode = BranchNode | ExtensionNode | LeafNode
 export type Nibbles = number[]
