+++ packages/trie/test/proof.spec.ts
@@ -1,6 +1,5 @@
 import tape from 'tape'
-import { CheckpointTrie } from '../src'
-import { LevelDB } from '../src/db'
+import { CheckpointTrie, LevelDB } from '../src'
 
 tape('simple merkle proofs generation and verification', function (tester) {
   const it = tester.test
